PDFChatApp/requirements.txt
langchain==0.0.319
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.27.2
faiss-cpu==1.7.4
streamlit-extras